Add files via upload
This is sample of n-gram model. We are going to implement winnowing algorithm. The first step is build a n-gram model to calculate the similarity of the input text. N-gram refers to the continuous occurrence of N words in a text. The n-element grammar model is a probabilistic language model based on the (N-1) order Markov chain, which deduces the sentence structure by the probability of the occurrence of N words. This model is widely used in the fields of probability theory, communication theory, computational linguistics (such as NLP based on statistics), computational biology (such as sequence analysis), and data compression. We are going to transfer similar text file to bytecode and use n-gram hash algorithm to analyze the bytecode and get the corresponding hash value. At last, the similarity is calculated by winnowing algorithm after extracting features.